feature/NS-50/managing-crates-to-be-pushed-to-cratesio: Split timescale-api into two parts
