+++ net-inserter-api/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "net-timescale-api"
+name = "net-inserter-api"
 version = "0.1.2"
 edition = "2021"
 description = "This crate defines types for the api used in the net-stalker project. Amazon Ion is used as the serialization format."
