+++ net-inserter-api/src/api/mod.rs
@@ -1,0 +1,3 @@
+//TODO: Rewrite all the inbound ion structs encode to `write_blob(*.encode())`
+
+pub mod network_packet;
