+++ net-reporter-api/NOTICE
@@ -1,0 +1,9 @@
+ion-rust
+Copyright 2021 Amazon.com, Inc. or its affiliates. All Rights Reserved. 
+
+---
+
+net-stalker Project:
+- Copyright (c) 2024 The net-stalker Developers
+- Licensed under the Apache License, Version 2.0
+  https://www.apache.org/licenses/LICENSE-2.0
