+++ net-reporter-api/README.md
@@ -1,0 +1,10 @@
+### License
+[license]: #license
+
+This crate is distributed under the terms of Apache License (Version 2.0).
+
+See [LICENSE](LICENSE) for details.
+
+All the copyright are mentioned.
+
+See [NOTICE](NOTICE)
