+++ net-reporter-api/src/api/mod.rs
@@ -3,10 +3,9 @@
 pub mod dashboard;
 
 //TODO: Add network prefix
-pub mod bandwidth_per_endpoint;
+pub mod network_bandwidth_per_endpoint;
 
 pub mod overview_dashboard_filters;
 
 pub mod network_bandwidth;
 pub mod network_graph;
-pub mod network_packet;
