+++ net-reporter-api/src/api/network_bandwidth/bandwidth_bucket.rs
