+++ net-reporter-api/src/api/network_bandwidth_per_endpoint/endpoint.rs
@@ -117,7 +117,7 @@ mod tests {
     use net_core_api::decoder_api::Decoder;
     use net_core_api::typed_api::Typed;
 
-    use crate::api::bandwidth_per_endpoint::endpoint::EndpointDTO;
+    use crate::api::network_bandwidth_per_endpoint::endpoint::EndpointDTO;
 
     #[test]
     fn test_correctly_read_encoded_endpoint() {
