+++ net-reporter-api/src/api/network_bandwidth_per_endpoint/mod.rs
@@ -1,0 +1,4 @@
+pub mod endpoint;
+#[allow(clippy::module_inception)]
+pub mod network_bandwidth_per_endpoint;
+pub mod network_bandwidth_per_endpoint_request;
