+++ net-reporter-api/src/api/network_graph/graph_node.rs
