--- net-timescale-api/src/api/bandwidth_per_endpoint/mod.rs
@@ -1,4 +1,0 @@
-pub mod endpoint;
-#[allow(clippy::module_inception)]
-pub mod bandwidth_per_endpoint;
-pub mod bandwidth_per_endpoint_request;
